fix: PR 리뷰 반영

+++ eatngo-core/src/main/kotlin/com/eatngo/product/infra/ProductPersistence.kt
@@ -8,5 +8,5 @@ interface ProductPersistence {
     fun findAllActivatedProductByStoreId(storeId: Long): List<Product>
     fun findAllActivatedProductsByStoreIds(storeIds: List<Long>): List<Product>
     fun findActivatedProductByIdAndStoreId(productId: Long, storeId: Long): Product?
-    fun countActiveProductsByStoreId(storeId: Long): Int
+    fun countActiveProductsByStoreId(storeId: Long): Long
 }
